requirements.txt
pillow
pandas
opencv-python
numba
numpy
pycocotools
scipy
pyclipper
shapely
imagesize
nltk
onnx
onnxruntime
onnxsim
tqdm
transformers
yapf
gpustat
calflops
gradio==5.49.1